Changed name of package logic to service Moved TaskConfigurationProperties to different location Added new Bean for test Task Repository
