+++ com.pgs.todo/src/main/java/com/pgs/todo/TaskConfigurationProperties.java
@@ -1,4 +1,4 @@
-package com.pgs.todo.model;
+package com.pgs.todo;
 
 import lombok.Getter;
 import lombok.Setter;
