+++ com.pgs.todo/src/test/java/com/pgs/todo/TestConfiguration.java
@@ -1,0 +1,60 @@
+package com.pgs.todo;
+
+import com.pgs.todo.model.Task;
+import com.pgs.todo.repository.TaskRepository;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+import java.util.*;
+
+@Configuration
+public class TestConfiguration {
+
+    @Bean
+    @Profile({"integration"})
+    TaskRepository testRepo() {
+        return new TaskRepository() {
+
+            private final Map<Integer, Task> tasks = new HashMap<>();
+
+            @Override
+            public List<Task> findAll() {
+                return new ArrayList<>(tasks.values());
+            }
+
+            @Override
+            public Page<Task> findAll(final Pageable page) {
+                return null;
+            }
+
+            @Override
+            public Optional<Task> findById(Integer id) {
+                return Optional.ofNullable(tasks.get(id));
+            }
+
+            @Override
+            public boolean existsById(Integer id) {
+                return tasks.containsKey(id);
+            }
+
+            @Override
+            public boolean existsByDoneIsFalseAndGroup_Id(final Integer groupId) {
+                return false;
+            }
+
+            @Override
+            public List<Task> findByDone(final boolean done){
+                return null;
+            }
+
+            @Override
+            public Task save(final Task entity) {
+                return tasks.put(tasks.size() + 1, entity);
+            }
+        };
+    }
+}
